remove data to js file
